requirements.txt
torchvision~=0.9.1
torch~=1.8.1
Pillow~=8.2.0
matplotlib~=3.3.4
numpy~=1.19.5